install_tinytex {tinytex}	R Documentation

Install/Uninstall TinyTeX

Description

The function install_tinytex() downloads the installation script from https://github.com/yihui/tinytex according to the platform (Unix or Windows), and executes it to install TinyTeX (a custom LaTeX distribution based on TeX Live). The function uninstall_tinytex() removes TinyTeX.

Usage

install_tinytex(force = FALSE, dir, repository = "ctan")

uninstall_tinytex(force = FALSE, dir = texlive_root())

Arguments

force
Whether to force to install (override) or uninstall TinyTeX.

dir
The directory to install or uninstall TinyTeX (should not exist unless force = TRUE).

repository
The CTAN repository to be used. By default, a fast mirror is automatically chosen. You can manually set one if the automatic mirror is not really fast enough, e.g., if you are in China, you may consider 'http://mirrors.tuna.tsinghua.edu.cn/CTAN/systems/texlive/tlnet'.

References

See the TinyTeX documentation (https://yihui.name/tinytex/) for the default installation directories on different platforms.

[Package tinytex version 0.4 Index]