is_tinytex {tinytex}	R Documentation

Check if the LaTeX installation is TinyTeX

Description

First find the root directory of the installation via tinytex_root(). Then check if the directory name is "tinytex" (case-insensitive). If not, further check if the first line of the file ‘texmf-dist/web2c/fmtutil.cnf’ under the directory contains "TinyTeX" or ".TinyTeX". If the binary version of TinyTeX was installed, ‘fmtutil.cnf’ should contain a line like ‘⁠Generated by */TinyTeX/bin/x86_64-darwin/tlmgr on Thu Sep 17 07:13:28 2020⁠’.

Usage

is_tinytex()

Value

A logical value indicating if the LaTeX installation is TinyTeX.

Examples

tinytex::is_tinytex()

[Package tinytex version 0.48 Index]