tl_sizes {tinytex}	R Documentation

Sizes of LaTeX packages in TeX Live

Description

Use the command tlmgr info --list to obtain the sizes of LaTeX packages.

Usage

tl_sizes(show_total = TRUE, pkgs = NULL, only_installed = TRUE, field = "size")

Arguments

show_total
Whether to show the total size.

pkgs
A character vector of package names (by default, all packages).

only_installed
Whether to list installed packages only.

field
A character vector of field names in the package information. See https://www.tug.org/texlive/doc/tlmgr.html#info for more info.

Value

By default, a data frame of three columns: package is the package names, size is the sizes in bytes, and size_h is the human-readable version of sizes. If different field names are provided in the field argument, the returned data frame will contain these columns.

[Package tinytex version 0.48 Index]