copy_tinytex {tinytex}	R Documentation

Copy TinyTeX to another location and use it in another system

Description

The function copy_tinytex() copies the existing TinyTeX installation to another directory (e.g., a portable device like a USB stick). The function use_tinytex() runs tlmgr path add to add the copy of TinyTeX in an existing folder to the PATH variable of the current system, so that you can use utilities such as tlmgr and pdflatex, etc.

Usage

copy_tinytex(
  from = tinytex_root(),
  to = select_dir("Select Destination Directory")
)

use_tinytex(from = select_dir("Select TinyTeX Directory"))

Arguments

from
The root directory of the TinyTeX installation. For copy_tinytex(), the default value tinytex_root() should be a reasonable guess if you installed TinyTeX via install_tinytex(). For use_tinytex(), if from is not provided, a dialog for choosing the directory interactively will pop up.

to
The destination directory where you want to make a copy of TinyTeX. Like from in use_tinytex(), a dialog will pop up if to is not provided in copy_tinytex().

Note

You can only copy TinyTeX and use it in the same system, e.g., the Windows version of TinyTeX only works on Windows.

[Package tinytex version 0.48 Index]