parse_packages {tinytex}	R Documentation

Find missing LaTeX packages from a LaTeX log file

Description

Analyze the error messages in a LaTeX log file to figure out the names of missing LaTeX packages that caused the errors. These packages can be installed via tlmgr_install(). Searching for missing packages is based on tlmgr_search().

Usage

parse_packages(
  log,
  text = readLines(log),
  files = detect_files(text),
  quiet = rep(FALSE, 3)
)

Arguments

log
Path to the LaTeX log file (typically named ‘*.log’).

text
A character vector of the error log (read from the file provided by the log argument by default).

files
A character vector of names of the missing files (automatically detected from the log by default).

quiet
Whether to suppress messages when finding packages. It should be a logical vector of length 3: the first element indicates whether to suppress the message when no missing LaTeX packages could be detected from the log, the second element indicate whether to suppress the message when searching for packages via tlmgr_search(), and the third element indicates whether to warn if no packages could be found via tlmgr_search().

Value

A character vector of LaTeX package names.

[Package tinytex version 0.34 Index]